+++ CHANGELOG.md
@@ -25,6 +25,7 @@ All notable changes to this project will be documented in this file.
 - *(ai)* Chess.com adapter and gemini ai working together
 - *(ai)* Companies_marketcap adapter and gemini ai working together
 - *(ai)* Countriesnow adapter and gemini ai working together
+- *(ai)* Devto adapter and gemini ai working together
 
 ### ⚙️ Miscellaneous Tasks